requirements.txt
numpy>=1.26
pandas>=2.0
scipy>=1.10
opencv-python>=4.8
scikit-image>=0.21
scikit-learn>=1.7.2
tqdm>=4.66
pyyaml>=6.0
pydantic>=2.6
pydantic-settings>=2.2
httpx
openai>=1.0.0
google-genai>=1.0.0
